CORE-7645 Refactor OntologyUtil tests
